requirements.txt
drevalpy>=1.4.0
torch>=2.0.0
pytorch-lightning>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
shap>=0.42.0
streamlit>=1.28.0
joblib>=1.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pyjwt>=2.8.0
